agent: increase max number of autoconfig attempts
Because https://jira.prplfoundation.org/browse/PPM-720 gets in the way
when we want to run wireless backhaul tests, increase the max number
of attempts for now.

Also change the type so that it can fit a higher number.

MAP-4.2.3_ETH_FH24G:turris-omnia
MAP-4.2.3_BHWIFI_FH24G:turris-omnia

+++ agent/src/beerocks/slave/tasks/ap_autoconfiguration_task.h
@@ -56,7 +56,7 @@ private:
 
     struct sStateStatus {
         eState state     = eState::UNCONFIGURED;
-        uint8_t attempts = 0;
+        uint16_t attempts = 0;
         std::chrono::steady_clock::time_point timeout;
     };
